requirements.txt
fastapi~=0.109.0
uvicorn~=0.25.0
tensorflow
pillow
keras


